Cast Postgres aggregates to numeric types; robust jsonb tag filter
Postgres returns COUNT/SUM(int) as bigint and AVG/SUM(numeric) as numeric,
which postgres-js hands back as strings — so display counts/averages typed
sql<number> (follower/project/vote/subscriber counts, rating and jam-score
averages, analytics event counts and durations) came back as strings, breaking
frontend math and recharts. Add ::int / ::float casts so they return real
numbers. Money sums stay sql<string> (decimal.js) and the delivery-bytes sum
stays a numeric string (fed straight into new Decimal()). Tag filter moves from
json_each to the null/scalar-safe @> containment operator.

+++ apps/api/src/routes/integrations.ts
@@ -33,32 +33,32 @@ const integrationRoutes = new Hono()
 
 		const [overview] = await db
 			.select({
-				totalEvents: sql<number>`COUNT(*)`,
-				totalDuration: sql<number>`COALESCE(SUM(duration_seconds), 0)`,
-				uniqueViewers: sql<number>`COUNT(DISTINCT user_id)`,
-				views: sql<number>`COUNT(*) FILTER (WHERE event_type = 'page_view')`,
-				plays: sql<number>`COUNT(*) FILTER (WHERE event_type = 'play')`,
-				watches: sql<number>`COUNT(*) FILTER (WHERE event_type = 'watch')`,
-				reads: sql<number>`COUNT(*) FILTER (WHERE event_type = 'read')`,
-				listens: sql<number>`COUNT(*) FILTER (WHERE event_type = 'listen')`,
+				totalEvents: sql<number>`COUNT(*)::int`,
+				totalDuration: sql<number>`COALESCE(SUM(duration_seconds), 0)::float`,
+				uniqueViewers: sql<number>`COUNT(DISTINCT user_id)::int`,
+				views: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'page_view'))::int`,
+				plays: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'play'))::int`,
+				watches: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'watch'))::int`,
+				reads: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'read'))::int`,
+				listens: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'listen'))::int`,
 			})
 			.from(attentionEvents)
 			.where(and(eq(attentionEvents.creatorId, user.id), gte(attentionEvents.createdAt, since)));
 
 		// Content counts
 		const [projectCount] = await db
-			.select({ count: sql<number>`COUNT(*)` })
+			.select({ count: sql<number>`COUNT(*)::int` })
 			.from(projects)
 			.where(eq(projects.creatorId, user.id));
 
 		const [postCount] = await db
-			.select({ count: sql<number>`COUNT(*)` })
+			.select({ count: sql<number>`COUNT(*)::int` })
 			.from(posts)
 			.where(eq(posts.creatorId, user.id));
 
 		// Cross-publish stats
 		const [publishCount] = await db
-			.select({ count: sql<number>`COUNT(*)` })
+			.select({ count: sql<number>`COUNT(*)::int` })
 			.from(crossPublishResults)
 			.where(eq(crossPublishResults.userId, user.id));
 
@@ -96,8 +96,8 @@ const integrationRoutes = new Hono()
 					projectId: attentionEvents.projectId,
 					projectTitle: projects.title,
 					projectSlug: projects.slug,
-					eventCount: sql<number>`COUNT(*)`,
-					totalDuration: sql<number>`COALESCE(SUM(${attentionEvents.durationSeconds}), 0)`,
+					eventCount: sql<number>`COUNT(*)::int`,
+					totalDuration: sql<number>`COALESCE(SUM(${attentionEvents.durationSeconds}), 0)::float`,
 				})
 				.from(attentionEvents)
 				.innerJoin(projects, eq(attentionEvents.projectId, projects.id))
@@ -129,8 +129,8 @@ const integrationRoutes = new Hono()
 				.select({
 					postId: attentionEvents.postId,
 					postTitle: posts.title,
-					eventCount: sql<number>`COUNT(*)`,
-					totalDuration: sql<number>`COALESCE(SUM(${attentionEvents.durationSeconds}), 0)`,
+					eventCount: sql<number>`COUNT(*)::int`,
+					totalDuration: sql<number>`COALESCE(SUM(${attentionEvents.durationSeconds}), 0)::float`,
 				})
 				.from(attentionEvents)
 				.innerJoin(posts, eq(attentionEvents.postId, posts.id))
@@ -169,11 +169,11 @@ const integrationRoutes = new Hono()
 		const timeseries = await db
 			.select({
 				date: dateExpr,
-				views: sql<number>`COUNT(*) FILTER (WHERE event_type = 'page_view')`,
-				plays: sql<number>`COUNT(*) FILTER (WHERE event_type = 'play')`,
-				watches: sql<number>`COUNT(*) FILTER (WHERE event_type = 'watch')`,
-				reads: sql<number>`COUNT(*) FILTER (WHERE event_type = 'read')`,
-				listens: sql<number>`COUNT(*) FILTER (WHERE event_type = 'listen')`,
+				views: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'page_view'))::int`,
+				plays: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'play'))::int`,
+				watches: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'watch'))::int`,
+				reads: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'read'))::int`,
+				listens: sql<number>`(COUNT(*) FILTER (WHERE event_type = 'listen'))::int`,
 			})
 			.from(attentionEvents)
 			.where(and(eq(attentionEvents.creatorId, user.id), gte(attentionEvents.createdAt, since)))
